Update
Making it so I can dynamically change the colors of the exr. First solution was to change the hue but that didn't look good and i realized instead of having the exr be in grey scale and changing the hue I can make it two separate colors then change those colors specifically to what i want. Took a little bit of creative thinking but I got there. Also added new exr and have it cycle through them.