Update
Making server action see the images as an array calling it in the star-scene then displaying it in a vector point path and animating through them based on scroll wheel event. Some small fixes too with folder paths and just small stuff in general

Challenges: remembering to always keep it in the foreach format because it is an array. Figuring out that I just need a single const stored to get the float between 1-0 and plugging that into the time based on the scroll.